Unimplement get/getAsX.
Methods were originally included for symmetry with built-in Optional,
but in retrospect that was a mistake. This already isn't a drop-in
replacement, and the use of get/getAsX on Optional is already highly
discouraged in favor of orElseThrow. Since we have orElseThrow as
well, get/getAsX is unnecessary (and since no one else is using this
library I don't have to worry as much about backwards compatibility)